apps/api/pyproject.toml
[project]
name = "moose-api"
version = "0.1.0"
description = "Moose Sports Empire - Fantasy Baseball API"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "arq>=0.26.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "cryptography>=44.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.18",
    "markdown>=3.7",
    "bleach>=6.2.0",
    "google-genai>=1.0.0",
    "numpy>=2.2.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.5",
    "pybaseball>=2.2.0",
    "rapidfuzz>=3.11.0",
    "baseball-id>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "anyio[trio]>=4.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008", "E402"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/moose_api"]
